Hide text field label if aria-label passed
When displaying input in a table we need to show validation
errors on that field, and ui/forms/field component is the one that
contains that logic. So, we have to use it instead of ui/forms/input,
and this change is needed to avoid showing redundant labels above
inputs (since we already have fields names in the table header). To be
a11y compliant we pass aria-label parameter for the input. If a
string is passed it will be used as aria-label value, otherwise (e.g.
`true`) - default label for the field will be assigned.